
\name{ContingentValuation-internal}
\alias{DBCVest}
\alias{SBCVest}
\title{Internal ContingentValuation functions 
}
\description{
Internal Contingent Valuation functions
}

\details{
  These are not to be called by the user (or in some cases are just
  waiting for proper documentation to be written :).
}
\keyword{ internal }




