\name{ContingentValuation-package}
\alias{ContingentValuation-package}
\alias{ContingentValuation}
\docType{package}
\title{
Package estimates coefficients from contingent valuation studies and calculates willingness to pay.
}
\description{
This Package will be helpful after you have collected your data in a contingent valuation study and need to estimate the coefficients and afterwards want to calculate the estimated willingness to pay. 
}
\details{
\tabular{ll}{
Package: \tab ContingentValuation\cr
Type: \tab Package\cr
Version: \tab 0.0.3\cr
Date: \tab 2012-11-30\cr
License: \tab GPL-2\cr
}
The Package currently contains a function to estimate a probit single and double bounded contingent valuation model and a function to calculate the willingness to pay. More is planned for the future.}
\author{
Ulrich B. Morawetz

Maintainer: Morawetz, U.B. <ulrich.morawetz@boku.ac.at>
}
\references{
Haab, T.C. and McConnell, K.E. (2003), \emph{Valuing Environmental and Natural Resources. The Econometrics of non-market Valuation.} Cheltenham, UK: Edward Elgar 

Morawetz, U.B., DeGroote, H. and Kimenju, S.C. (2011) Improving the Use of Experimental Auctions in Africa: Theory and Evidence, \emph{Journal of Agricultural and Resource Economics}, 36(2):263-279.

}
\keyword{Contingent Valuation}
%\seealso{
%}

%\examples{
%No expamles
%}

